re: #346 Added the choice of using either center-of-gravity (COG) or peak start. Also added handling of cases where scanlines are invalid, or cases where RANSAC is unable to estimate line parameters (WIP).

+++ PlusLib/src/CalibrationAlgo/TemporalCalibrationAlgo.h
@@ -140,7 +140,7 @@ private:
   void plotIntArray(std::vector<int> intensityValues);
   void plotDoubleArray(std::vector<double> intensityValues);
   PlusStatus FindPeakStart(std::vector<int> &intensityProfile,int MaxFromLargestArea,
-                           int startOfMaxArea, int &startOfPeak);
+                           int startOfMaxArea, double &startOfPeak);
   PlusStatus FindLargestPeak(std::vector<int> &intensityProfile,int &MaxFromLargestArea,
                                                 int &MaxFromLargestAreaIndex, int &startOfMaxArea);
   PlusStatus ComputeCenterOfGravity(std::vector<int> &intensityProfile, int startOfMaxArea,
